Also rename viewer_page dir to timeline_page
