+++ ui/src/frontend/help_modal.ts
@@ -23,7 +23,7 @@ import {
   nativeKeyboardLayoutMap,
   NotSupportedError,
 } from '../base/keyboard_layout_map';
-import {KeyMapping} from './viewer_page/wasd_navigation_handler';
+import {KeyMapping} from './timeline_page/wasd_navigation_handler';
 import {raf} from '../core/raf_scheduler';
 
 export function toggleHelp() {
