+++ ui/src/frontend/timeline_page/current_selection_tab.ts
