+++ ui/src/frontend/timeline_page/gridline_helper.ts
