+++ ui/src/frontend/timeline_page/minimap.ts
