+++ ui/src/frontend/timeline_page/notes_panel.ts
