+++ ui/src/frontend/timeline_page/resolution.ts
