+++ ui/src/frontend/timeline_page/tab_panel.ts
