+++ ui/src/frontend/timeline_page/time_axis_panel.ts
