+++ ui/src/frontend/timeline_page/time_selection_panel.ts
