+++ ui/src/frontend/timeline_page/timeline_interactions.ts
