+++ ui/src/frontend/timeline_page/track_view.ts
