+++ ui/src/frontend/timeline_page/wasd_navigation_handler.ts
